Now widgets can just update their drawstate at will
Still not sure if I like it, but that at least will work without
too much hassle.

the 'update()' idea can be cumbersome once the update logic becomes
more complex. Also we don't have any of the things instantiated yet
that we would want to update.

The only solution is stateful, but that makes updating based on
state changes hard. So rather have one algorithm and rebuild the
whole thing, instead of surgically update pieces of the state.