Move ownership of marked entries to the MarkPane
interesting experience... it's easier to just keep state central, and
have one place where mutation happens. Now that things are all over
the place, it becomes more difficult to handle.
However, I believe there is a way, the problem here is the
hybrid state the program is currenlty in... partly centralized, partly
localized.