Pune/India: Rustic way of handling the mark panes disappearance
Interesting: before that fix, the GUI was in an invalid state,
as the mark pane disappeared without setting the focus back to the
main window.
Thanks to everything being encoded in the typesystem and no assumptions
made, the program wouldn't crash when buttons are pressed, which would
otherwise be attempted to dispatch to the now gone mark pane.

Instead one could press 'q' or tab, as fortunately, the pane
handling is always done before any of the panes get a chance.