Handle broken symlinks, they can now be deleted
These are tough, as metadata on them will fail, as it tries to follow
the symlink. It's odd, as we can still check if this was a symlink
in case it could be followed, but we can't check that in case it could
not be followed. In a way, the `metadata()` call should rather amended
with a sybling method like `nmetadata()` to have a version that doesn't
follow symlinks.
Anyway... .