Fix color scheme for light terminals
This is done by making sure that we never have a background that is
black or white combined with a foreground that uses the terminal's
default and vice versa.

Because we cannot access the default terminal colors, we have to rely on
inverted colors and bold text for highlighting.

Also, the mark pane was improved to be more consistent with the entries
pane. Ideally, we would use the same color as in the entries pane, but
this is currently not possible, because the mark pane does not know
whether a path is a directory or not.

Fixes #13.