Use full path for obtaining the 'real size on disk'
This fixes a long standing issue on windows, #53, which relies
on the full path to be used.

On linux, it seems to be enough to use the meta-data, which is also
passed along.
On windows, the path seems to actually be used, explaining why it
wonderously didn't work on windows before.