feat: interactive mode learns 'toggle [a]ll' and 'remove [a]ll'.
In the mark pane, the 'a' key will now toggle all entries.
This is particularly interesting for selecting entries to
exclude by hande and then invert the selection by toggling [a]ll.

In the mark pane, toggling all with the 'a' key means removing
all entries and closing the pane.