fix: keep checking for existance of entries outside of the glob top-level.
The glob top-level is used to display all search results which means
that there can be a lot of them, which would unnecessarily slow down
the search operation.

Previously it would never check for the existence of an entry in glob mode,
but now it will do so outside of the top-level.