feat: add `--log-file` flag to keep track of some debug info, which includes panics.
Previously, when `dua i` was used, panics would be hard to observe, if at all,
as they would print to the alternate screen. Now, when the `--log-file dua.log`
is specified, the panic will be emitted into the log file instead and thus won't
be lost anymore.

This may help with debugging in future.