Replace tui_react::Terminal with tui::Terminal
This is an experiment in converting a program that used the Terminal
provided by tui-react into a program that uses the Terminal provided by
ratatui. My hypothesis was that the two were not significantly different
from eachother, and this seems to be true in this case.

I found that `Widget` trait used by ratatui's `Terminal` is sufficiently
powerful to pass arbitrary mutable state to rendering code. A relatively
simple `FunctionWidget` trampoline is used to achieve this.

The tui_react dependency is retained. The code still calls some
tui_react utility functions and uses its List widget instead of the
ratatui List widget. These utilities do not depend on the tui_react
Terminal and could, in theory, be split out of tui-react into a separate
crate.